Unify buzz methods as MarlinUI::buzz (#14803)
